requirements.txt
# Deep Learning Framework
torch>=1.9.0
torchvision>=0.10.0

# Image Processing
Pillow>=8.0.0

# Scientific Computing
numpy>=1.21.0

# Data Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Utilities
argparse
